move compare row to common.
